BUG: Fix numpy boolean subtraction error in Series.diff (#27755)
